Move form component into components folder and import into signup.js page component
