+++ src/components/signupForm/styledComponents.js
@@ -1,3 +1,4 @@
+import React from "react"
 import styled from "styled-components"
 import DialogTitle from "@material-ui/core/DialogTitle"
 import Button from "@material-ui/core/Button"
