+++ src/pages/signup.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Form from "./form"
+import Form from "../components/signupForm"
 
 export default () => {
   return <Form />
